rename factories -> blog_factory
