+++ apps/core/management/commands/filldb.py
@@ -2,7 +2,7 @@ from typing import Any, Optional
 
 from django.core.management.base import BaseCommand, CommandError
 
-from apps.articles.tests.factories import BlogFactory
+from apps.articles.tests.blog_factory import BlogFactory
 from apps.core.tests.factories import ImageFactory, PersonFactory, UserFactory
 from apps.info.tests.factories import (
     FestivalFactory,
